work

+++ src/main/java/com/back/domain/post/post/controller/ApiV1PostController.java
@@ -1,8 +1,11 @@
 package com.back.domain.post.post.controller;
 
+import com.back.domain.member.member.entity.Member;
 import com.back.domain.post.post.dto.PostDto;
 import com.back.domain.post.post.entity.Post;
 import com.back.domain.post.post.service.PostService;
+import com.back.global.exception.ServiceException;
+import com.back.global.rq.Rq;
 import com.back.global.rsData.RsData;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -11,6 +14,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,9 +22,12 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/v1/posts")
 @RequiredArgsConstructor
+@Validated
 @Tag(name = "ApiV1PostController", description = "API 글 컨트롤러")
 public class ApiV1PostController {
     private final PostService postService;
+    private final Rq rq;
+
 
     @GetMapping
     @Transactional(readOnly = true)
@@ -46,8 +53,15 @@ public class ApiV1PostController {
     @DeleteMapping("/{id}")
     @Transactional
     @Operation(summary = "삭제")
-    public RsData<Void> delete(@PathVariable int id) {
+    public RsData<Void> delete(
+            @PathVariable int id
+    ) {
+        Member actor = rq.getActor();
         Post post = postService.findById(id).get();
+
+        if(!actor.equals(post.getMember())) {
+            throw new ServiceException("403-1", "권한이 없습니다.");
+        }
 
         postService.delete(post);
 
@@ -65,14 +79,22 @@ public class ApiV1PostController {
             @NotBlank
             @Size(min = 2, max = 5000)
             String content
+
     ) {
     }
 
     @PostMapping
     @Transactional
     @Operation(summary = "작성")
-    public RsData<PostDto> write(@Valid @RequestBody PostWriteReqBody reqBody) {
-        Post post = postService.write(reqBody.title, reqBody.content);
+    public RsData<PostDto> write(
+            @Valid @RequestBody PostWriteReqBody reqBody
+
+    ) {
+        Member actor = rq.getActor();
+
+        Post post = postService.write(actor, reqBody.title, reqBody.content);
+
+
 
         return new RsData<>(
                 "201-1",
@@ -97,8 +119,17 @@ public class ApiV1PostController {
     public RsData<Void> modify(
             @PathVariable int id,
             @Valid @RequestBody PostModifyReqBody reqBody
+
+
     ) {
+        Member actor = rq.getActor();
+
         Post post = postService.findById(id).get();
+
+        if(!actor.equals(post.getMember())) {
+            throw new ServiceException("403-1", "권한이 없습니다.");
+        }
+
         postService.modify(post, reqBody.title, reqBody.content);
 
         return new RsData<>(
